Added functionality to Pleb's "Remind" command
Remind command is now a working command, but still does not work when
asking the bot to remind someone else to do something. The if statement
on line 35 functions correctly, however the code on line 37 must be
updated from msg.reply to a different form of msg. Afaik msg.reply will
always start by mentioning the issuer of the original command.